change internal structure of redux components
